requirements.txt
groq
python-dotenv
numpy
pandas
matplotlib
tabulate
python-docx
pypdf